Renaming, cleanup.
