+++ src/main/scala/net/ripe/rpki/publicationserver/messaging/Messages.scala
@@ -5,17 +5,17 @@ import java.util.UUID
 
 import net.ripe.rpki.publicationserver.QueryMessage
 import net.ripe.rpki.publicationserver.model.ClientId
-import net.ripe.rpki.publicationserver.store.XodusObjectStore
+import net.ripe.rpki.publicationserver.store.ObjectStore
 
 object Messages {
 
   case class RawMessage(message: net.ripe.rpki.publicationserver.Message, clientId: ClientId)
 
-  case class ValidatedMessage(queryMessage: QueryMessage, state: XodusObjectStore.State)
+  case class ValidatedMessage(queryMessage: QueryMessage, state: ObjectStore.State)
 
-  case class BatchMessage(messages: Seq[QueryMessage], state: XodusObjectStore.State)
+  case class BatchMessage(messages: Seq[QueryMessage], state: ObjectStore.State)
 
-  case class InitRepo(state: XodusObjectStore.State)
+  case class InitRepo(state: ObjectStore.State)
 
   case class CleanUpSnapshot(timestamp: FileTime, serial: Long)
