+++ src/test/scala/net/ripe/rpki/publicationserver/RepositoryStateTest.scala
@@ -20,7 +20,7 @@ object RepositoryStateTest {
 
   var retainPeriodOverride: Int = 1
 
-  val theObjectStore = XodusObjectStore.get
+  val theObjectStore = ObjectStore.get
   val theRsyncWriter = MockitoSugar.mock[RsyncRepositoryWriter]
 
   lazy val conf = new AppConfig {
